service/requirements.txt
Flask == 0.12.2
flask_cors == 3.0.3
biopython == 1.70

